fix: extact enum to upper level
